requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.6.0
numpy>=1.26.0

# Relational store for the insurance tools (SQLite via SQLAlchemy).
sqlalchemy>=2.0.0

# LangChain stack
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.7
langchain-community>=0.2.0

# Keep the RAG pure-Python (no FAISS) — uses an in-memory vector store
# driven by OpenAI embeddings.
